+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Scripts.hs
@@ -336,7 +336,6 @@ instance
   EncCBORGroup (AlonzoPlutusPurpose f era)
   where
   listLen _ = 2
-  listLenBound _ = 2
   encCBORGroup = \case
     AlonzoSpending p -> encodeWord8 0 <> encCBOR p
     AlonzoMinting p -> encodeWord8 1 <> encCBOR p